test: separate the user tests
Before, all user tests were in a single file. Each feature test now has its own file.
